Add support for both fixed- and unbounded- vector types in interpreter
This commit adds basic support for `%fvec(...)` and `%vec(...)` types,
the former being expecially important as it is used to model register
banks.

For now, both are implemented using Z3's `McCarthyArray`. This is
probably not the best idea for fixed-size vectors but will do for now.

This commit also adds `vector_access` and `vector_update` externs.