Add support for "unknown-length" bitvector type in the interpreter
This commit adds basic support "unknown-length" bitrector type `%bv`.
Such type is represented as Z3 node of uninterpreted sort. Such node is
really only needed before first assignment to variable `Instr_Decl`d
as `%bv` - the assigned value will know it's length. This is just fine,
if the program is not well-formed, it should not reach the interpreter
at all.