Debugger: improve UX when stepping over/into `goto` and `end` instrs
In Jib, `goto` and `end` instructions cannot have a source location. So
when Jib debugger steps on such instruction, it stops showing Sail
source and shows only instructions. This is bit annoying especially
for `goto` as with another step, Sail source is shown again.

To somewhat improve this, change `JibInterpreterContext >> #sourceLoc`
to return location of instruction preceding `goto` and `end`.